requirements/requirements.txt
# Copyright 2023 Google LLC.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

absl-py
arviz==0.11.2
immutabledict>=2.0.0
jax>=0.3.18
jaxlib>=0.3.18
matplotlib==3.6.1
numpy>=1.24
numpyro>=0.9.2
pandas>=1.1.5
scipy
seaborn==0.11.1
scikit-learn
statsmodels>=0.13.0
tensorflow>=2.7.2
typing_extensions;python_version<'3.8'